Working Struct

+++ src/Cpp/lib/Socket.h
@@ -30,7 +30,7 @@ public:
 };
 
 struct SimpStruct{
-    float a;
-    float b;
-    int c;
+    int a;
+    int b;
+    long c;
 };
